+++ Changelog.md
@@ -2,8 +2,8 @@
 
 ### Added
 
--   J-Link devices OB FW can now be automatically updated when erasing, writing, and
-    reading.
+-   J-Link devices OB FW can now be automatically updated when erasing, writing,
+    and reading.
 -   Button to manually update the OB Firmware of a device.
 
 ### Changed